fix state change detection
