+++ pkg/nmlite/interface.go
@@ -683,116 +683,6 @@ func (im *InterfaceManager) monitorInterfaceState() {
 
 }
 
-// updateInterfaceState updates the current interface state
-func (im *InterfaceManager) updateInterfaceState() error {
-	nl, err := im.link()
-	if err != nil {
-		return fmt.Errorf("failed to get interface: %w", err)
-	}
-
-	// Check if state changed
-	stateChanged := false
-
-	attrs := nl.Attrs()
-	isUp := attrs.OperState == netlink.OperUp
-
-	// check if the interface has unicast addresses
-	isOnline := isUp && nl.HasGlobalUnicastAddress()
-
-	// We should release the lock before calling the callbacks
-	// to avoid deadlocks
-	im.stateMu.Lock()
-	if im.state.Up != isUp {
-		im.state.Up = isUp
-		stateChanged = true
-	}
-
-	if im.state.Online != isOnline {
-		im.state.Online = isOnline
-		stateChanged = true
-	}
-
-	if im.state.MACAddress != attrs.HardwareAddr.String() {
-		im.state.MACAddress = attrs.HardwareAddr.String()
-		stateChanged = true
-	}
-
-	// Update IP addresses
-	if err := im.updateIPAddresses(nl); err != nil {
-		im.logger.Error().Err(err).Msg("failed to update IP addresses")
-	}
-
-	im.state.LastUpdated = time.Now()
-	im.stateMu.Unlock()
-
-	// Notify callback if state changed
-	if stateChanged && im.onStateChange != nil {
-		im.logger.Debug().Interface("state", im.state).Msg("notifying state change")
-		im.onStateChange(*im.state)
-	}
-
-	return nil
-}
-
-// updateIPAddresses updates the IP addresses in the state
-func (im *InterfaceManager) updateIPAddresses(nl *link.Link) error {
-	if err := nl.Refresh(); err != nil {
-		return fmt.Errorf("failed to refresh link: %w", err)
-	}
-
-	addrs, err := nl.AddrList(link.AfUnspec)
-	if err != nil {
-		return fmt.Errorf("failed to get addresses: %w", err)
-	}
-
-	var (
-		ipv4Addresses        []string
-		ipv6Addresses        []types.IPv6Address
-		ipv4Addr, ipv6Addr   string
-		ipv6LinkLocal        string
-		ipv4Ready, ipv6Ready = false, false
-	)
-
-	for _, addr := range addrs {
-		im.logger.Debug().
-			IPAddr("address", addr.IP).
-			Msg("checking address")
-		if addr.IP.To4() != nil {
-			// IPv4 address
-			ipv4Addresses = append(ipv4Addresses, addr.IPNet.String())
-			if ipv4Addr == "" {
-				ipv4Addr = addr.IP.String()
-				ipv4Ready = true
-			}
-		} else if addr.IP.To16() != nil {
-			// IPv6 address
-			if addr.IP.IsLinkLocalUnicast() {
-				ipv6LinkLocal = addr.IP.String()
-			} else if addr.IP.IsGlobalUnicast() {
-				ipv6Addresses = append(ipv6Addresses, types.IPv6Address{
-					Address: addr.IP,
-					Prefix:  *addr.IPNet,
-					Scope:   addr.Scope,
-				})
-				if ipv6Addr == "" {
-					ipv6Addr = addr.IP.String()
-					ipv6Ready = true
-				}
-			}
-		}
-	}
-
-	im.state.IPv4Addresses = ipv4Addresses
-	im.state.IPv6Addresses = ipv6Addresses
-	im.state.IPv6LinkLocal = ipv6LinkLocal
-	im.state.IPv4Address = ipv4Addr
-	im.state.IPv6Address = ipv6Addr
-	im.state.IPv4Ready = ipv4Ready
-	im.state.IPv6Ready = ipv6Ready
-
-	return nil
-}
-
 // updateStateFromDHCPLease updates the state from a DHCP lease
 func (im *InterfaceManager) updateStateFromDHCPLease(lease *types.DHCPLease) {
 	im.stateMu.Lock()
